refactor: Rename MockDB into MockUserDB
